feat: Introduce NativeToken and TokenBalances composables (#296)
* feat: Introduce NativeToken and TokenBalances composables

* refactor: Introduce useStaking

* refactor: Split Transaction Logic into it's own composable

* feat: limit retries to tx history, per #293
