+++ src/composables/index.ts
@@ -1,13 +1,21 @@
-import useHomeModal from './useHomeModal'
-import useRadix from './useRadix'
-import useSidebar from './useSidebar'
-import useWallet from './useWallet'
 import useConnectableRadix from './useConnectableRadix'
+import useHomeModal from './useHomeModal'
+import useNativeToken from './useNativeToken'
+import useRadix from './useRadix'
+import useStaking from './useStaking'
+import useSidebar from './useSidebar'
+import useTransactions from './useTransactions'
+import useTokenBalances from './useTokenBalances'
+import useWallet from './useWallet'
 
 export {
+  useConnectableRadix,
   useHomeModal,
+  useNativeToken,
   useRadix,
+  useStaking,
   useSidebar,
-  useWallet,
-  useConnectableRadix
+  useTransactions,
+  useTokenBalances,
+  useWallet
 }
