+++ src/views/Wallet/WalletStaking.vue
@@ -109,8 +109,8 @@ import FormField from '@/components/FormField.vue'
 import ButtonSubmit from '@/components/ButtonSubmit.vue'
 import { asBigNumber } from '@/components/BigAmount.vue'
 import { Position } from '@/services/_types'
-import { useRadix, useWallet } from '@/composables'
-import { useRouter } from 'vue-router'
+import { useNativeToken, useRadix, useStaking, useTransactions, useTokenBalances, useWallet } from '@/composables'
+import { useRouter, onBeforeRouteLeave } from 'vue-router'
 import { useI18n } from 'vue-i18n'
 
 interface StakeForm {
@@ -136,14 +136,27 @@ const WalletStaking = defineComponent({
     const router = useRouter()
     const {
       activeAddress,
-      tokenBalances,
-      nativeToken,
-      activeStakes,
-      activeUnstakes,
-      nativeTokenBalance,
-      stakeTokens,
-      unstakeTokens
+      activeAccount,
+      hardwareAccount,
+      hardwareAccountFailedToSign
     } = useWallet(radix, router)
+
+    const { stakeTokens, unstakeTokens, transactionUnsub } = useTransactions(radix, router, activeAccount.value, hardwareAccount.value, {
+      ledgerSigningError: () => {
+        hardwareAccountFailedToSign()
+      }
+    })
+
+    const { nativeToken, nativeTokenUnsub } = useNativeToken(radix)
+    const { tokenBalances, tokenBalanceFor, tokenBalancesUnsub } = useTokenBalances(radix)
+    const { activeStakes, activeUnstakes, stakingUnsub } = useStaking(radix)
+
+    onBeforeRouteLeave(() => {
+      nativeTokenUnsub()
+      tokenBalancesUnsub()
+      stakingUnsub()
+      transactionUnsub()
+    })
 
     const setForm = (form: string) => {
       activeForm.value = form
@@ -151,8 +164,11 @@ const WalletStaking = defineComponent({
     }
 
     // Computed Values
-    const xrdBalance: ComputedRef<AmountT> = computed(() =>
-      nativeTokenBalance.value ? nativeTokenBalance.value.amount : Amount.fromUnsafe(0)._unsafeUnwrap())
+    const xrdBalance: ComputedRef<AmountT> = computed(() => {
+      if (!tokenBalances.value || !nativeToken.value) return Amount.fromUnsafe(0)._unsafeUnwrap()
+      const nativeTokenBalance = tokenBalanceFor(nativeToken.value)
+      return nativeTokenBalance ? nativeTokenBalance.amount : Amount.fromUnsafe(0)._unsafeUnwrap()
+    })
 
     const sortedPositions: ComputedRef<Array<Position>> = computed(() => {
       if (!activeStakes.value || !activeUnstakes.value) return []
@@ -216,33 +232,32 @@ const WalletStaking = defineComponent({
     }
 
     const handleSubmitStake = () => {
-      if (meta.value.valid && nativeTokenBalance.value) {
-        const safeAddress = safelyUnwrapValidator(values.validator)
-        const safeAmount = safelyUnwrapAmount(Number(values.amount))
-        const greaterThanZero = safeAmount && validateGreaterThanZero(safeAmount)
-        const validAmount = safeAmount && validateAmountOfType(safeAmount, nativeTokenBalance.value.token)
+      if (!tokenBalances.value || !nativeToken.value) return
+      const nativeTokenBalance = tokenBalanceFor(nativeToken.value)
+      if (!meta.value.valid || !nativeTokenBalance) return
+      const safeAddress = safelyUnwrapValidator(values.validator)
+      const safeAmount = safelyUnwrapAmount(Number(values.amount))
+      const greaterThanZero = safeAmount && validateGreaterThanZero(safeAmount)
+      const validAmount = safeAmount && validateAmountOfType(safeAmount, nativeTokenBalance.token)
 
-        if (!greaterThanZero) {
-          setErrors({
-            amount: t('validations.greaterThanZero')
-          })
-        } else if (!validAmount) {
-          setErrors({
-            amount: t('validations.amountOfType', { granularity: nativeTokenBalance.value.token.granularity.toString() })
-          })
-        } else {
-          if (!safeAddress || !safeAmount) return
-          activeForm.value === 'stake'
-            ? stakeTokens({
-              validator: safeAddress,
-              amount: safeAmount
-            })
-            : unstakeTokens({
-              validator: safeAddress,
-              amount: safeAmount
-            })
-        }
+      if (!greaterThanZero) {
+        setErrors({ amount: t('validations.greaterThanZero') })
+        return
       }
+      if (!validAmount) {
+        setErrors({ amount: t('validations.amountOfType', { granularity: nativeTokenBalance.token.granularity.toString() }) })
+        return
+      }
+      if (!safeAddress || !safeAmount) return
+      activeForm.value === 'stake'
+        ? stakeTokens({
+          validator: safeAddress,
+          amount: safeAmount
+        })
+        : unstakeTokens({
+          validator: safeAddress,
+          amount: safeAmount
+        })
     }
 
     return {
